pipeline/requirements.txt
neo4j>=5.0.0
requests>=2.32.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
langchain-google-genai>=2.0.0
